perf(web): 首屏再瘦身 —— markdown/重页面懒加载、依赖块归位、Brotli 预压缩
在 #51 基础上继续压缩首屏关键路径（gzip 后 650KB → brotli 后 ~367KB）：

1. Markdown 渲染链懒加载：react-markdown + micromark/hast 生态 +
   highlight.js（约 332KB）从 vendor 拆出独立 markdown 块，
   Markdown 组件改懒加载壳（chunk 到位前按纯文本展示，不闪加载态），
   终端首屏不再为对话富文本买单。

2. 重页面按路由懒加载：Swarm / GitPanel / BrowserView / PhoneView
   切到对应 tab 才拉 chunk，index 块 331KB → 247KB；
   dnd 块（@hello-pangea/dnd + react-redux，99KB）只有 Swarm 用，
   随之自动退出首屏。

3. 依赖块归位：jszip(94KB)→office、lodash-es/es-toolkit→mermaid、
   state-local→monaco、react-redux 等→dnd。此前这些懒加载块的
   专属依赖都错落在首屏 vendor 里白白下载，vendor 563KB → 74KB。

4. Brotli 预压缩：构建期在 .gz 之外再产出 .br（quality 11，比 gzip
   再小 15~20%），后端按 Accept-Encoding 择优下发，gzip 兜底。

验证：headless Chrome CDP 冒烟——登录后逐一切换懒加载路由，
页面渲染正常、chunk 按需加载、无控制台报错；br 下发内容与原文件
逐字节一致；304/路径穿越/无压缩客户端行为不变。
